RavenDB-26838 Move CDC Sink SNMP test from RavenDB-21192 to RavenDB-26838
