+++ test/SlowTests.Issues/Issues/RavenDB-21192.cs
@@ -1707,133 +1707,6 @@ public class RavenDB_21192 : RavenTestBase
         }
     }
 
-    [RavenFact(RavenTestCategory.Monitoring | RavenTestCategory.Sinks)]
-    public async Task CanGetCdcSinkErrorsSnmpMetrics_V2C()
-    {
-        var port = ReservePort().Port;
-        var communityString = "public-test";
-        var customSettings = new Dictionary<string, string>
-        {
-            [RavenConfiguration.GetKey(x => x.Monitoring.Snmp.Enabled)] = "true",
-            [RavenConfiguration.GetKey(x => x.Monitoring.Snmp.SupportedVersions)] = "V2C",
-            [RavenConfiguration.GetKey(x => x.Monitoring.Snmp.Port)] = port.ToString(),
-            [RavenConfiguration.GetKey(x => x.Monitoring.Snmp.Community)] = communityString
-        };
-
-        UseNewLocalServer(customSettings);
-
-        using (var store = GetDocumentStore(new Options { CreateDatabase = true }))
-        {
-            const string cdcSinkName = "CdcSink1";
-
-            var connectionString = new SqlConnectionString
-            {
-                Name = "cdc-cs",
-                FactoryName = "Microsoft.Data.SqlClient",
-                ConnectionString = "Server=localhost;Database=TestDb;User Id=sa;Password=pass;"
-            };
-            store.Maintenance.Send(new PutConnectionStringOperation<SqlConnectionString>(connectionString));
-
-            store.Maintenance.Send(new AddCdcSinkOperation(new CdcSinkConfiguration
-            {
-                Name = cdcSinkName,
-                ConnectionStringName = connectionString.Name,
-                Disabled = true,
-                Tables = new List<CdcSinkTableConfig>
-                {
-                    new CdcSinkTableConfig
-                    {
-                        CollectionName = "Orders",
-                        SourceTableSchema = "public",
-                        SourceTableName = "orders",
-                        Columns = new List<CdcColumnMapping> { new CdcColumnMapping { Column = "order_id", Name = "OrderId" } },
-                        PrimaryKeyColumns = new List<string> { "order_id" }
-                    }
-                }
-            }));
-
-            var database = GetDatabase(store.Database).GetAwaiter().GetResult();
-            var now = DateTime.UtcNow;
-
-            for (int i = 0; i < 2; i++)
-            {
-                database.TaskErrorsStorage.StoreProcessError(TaskCategory.CdcSink, new TaskProcessError
-                {
-                    CreatedAt = now,
-                    TaskName = cdcSinkName,
-                    Step = TaskErrorStep.Extraction,
-                    Error = "process error"
-                });
-            }
-
-            database.TaskErrorsStorage.StoreItemErrors(TaskCategory.CdcSink, cdcSinkName,
-            [
-                new TaskItemError { DocumentId = "orders/1", TaskName = cdcSinkName, CreatedAt = now, Step = TaskErrorStep.Load, Error = "e1" },
-                new TaskItemError { DocumentId = "orders/2", TaskName = cdcSinkName, CreatedAt = now, Step = TaskErrorStep.Load, Error = "e2" },
-                new TaskItemError { DocumentId = "orders/3", TaskName = cdcSinkName, CreatedAt = now, Step = TaskErrorStep.Transformation, Error = "e3" }
-            ]);
-
-            const int expectedErrors = 5;
-
-            var ip = new Uri(Server.WebUrl).Host;
-            var endpoint = new IPEndPoint(IPAddress.Parse(ip), port);
-
-            string serverErrorsOid = null, dbErrorsOid = null;
-            string taskErrorsOid = null, taskHealthOid = null, taskResponsibleNodeOid = null, taskLastBatchOid = null;
-
-            Assert.True(WaitForValue(() =>
-            {
-                using (var commands = store.Commands())
-                {
-                    var cmd = new GetSnmpOidsCommand();
-                    commands.Execute(cmd);
-
-                    if (cmd.Result is not BlittableJsonReaderObject res)
-                        return false;
-
-                    if (res.TryGet("Server", out BlittableJsonReaderArray serverEntries) == false ||
-                        res.TryGet("Databases", out BlittableJsonReaderObject databases) == false ||
-                        databases.TryGet(store.Database, out BlittableJsonReaderObject databaseOids) == false ||
-                        databaseOids.TryGet("@General", out BlittableJsonReaderArray generalEntries) == false ||
-                        databaseOids.TryGet("CdcSinks", out BlittableJsonReaderObject cdcSinks) == false ||
-                        cdcSinks.TryGet(cdcSinkName, out BlittableJsonReaderArray cdcSinkEntries) == false ||
-                        cdcSinkEntries == null)
-                        return false;
-
-                    string Find(BlittableJsonReaderArray entries, string description) =>
-                        JsonConvert.DeserializeObject<List<SnmpEntry>>(entries.ToString()).SingleOrDefault(x => x.Description == description)?.OID;
-
-                    serverErrorsOid = Find(serverEntries, "Number of CDC Sink errors");
-                    dbErrorsOid = Find(generalEntries, "Number of CDC Sink errors");
-                    taskErrorsOid = Find(cdcSinkEntries, "Number of CDC Sink task errors");
-                    taskHealthOid = Find(cdcSinkEntries, "Health status of particular CDC Sink task");
-                    taskResponsibleNodeOid = Find(cdcSinkEntries, "Responsible node tag of particular CDC Sink task");
-                    taskLastBatchOid = Find(cdcSinkEntries, "Last successful batch time");
-
-                    return serverErrorsOid != null && dbErrorsOid != null && taskErrorsOid != null &&
-                           taskHealthOid != null && taskResponsibleNodeOid != null && taskLastBatchOid != null;
-                }
-            }, true, timeout: 20000, interval: 500));
-
-            ISnmpData SnmpGet(string oid)
-            {
-                var result = Messenger.Get(VersionCode.V2, endpoint, new OctetString(communityString),
-                    [new Variable(new ObjectIdentifier(oid))], 10000);
-                return result.Single().Data;
-            }
-
-            Assert.Equal(expectedErrors, ((Integer32)SnmpGet(serverErrorsOid)).ToInt32());
-            Assert.Equal(expectedErrors, ((Integer32)SnmpGet(dbErrorsOid)).ToInt32());
-            Assert.Equal(expectedErrors, ((Integer32)SnmpGet(taskErrorsOid)).ToInt32());
-
-            Assert.Equal(nameof(EtlProcessHealthStatus.Healthy), SnmpGet(taskHealthOid).ToString());
-
-            Assert.Equal(Server.ServerStore.NodeTag, SnmpGet(taskResponsibleNodeOid).ToString());
-
-            Assert.Equal(SnmpType.TimeTicks, SnmpGet(taskLastBatchOid).TypeCode);
-        }
-    }
-
     [RavenFact(RavenTestCategory.Monitoring | RavenTestCategory.Etl)]
     public async Task EtlsMonitoringEndpointShouldWork()
     {
